% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_errors.R
\name{check_lifx_response}
\alias{check_lifx_response}
\title{react to 'LIFX' api response error codes}
\usage{
check_lifx_response(response)
}
\arguments{
\item{response}{the api response received from httr::PUT / POST / GET}
}
\value{
depending on the status either: an error; a warning and the response as is; the response as is without any message.
}
\description{
react to 'LIFX' api response error codes
}
\references{
error messages copied from \url{https://api.developer.lifx.com/docs/errors}
}
