% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifx_documentation.R
\docType{package}
\name{lifx}
\alias{lifx}
\title{lifx: A package for controlling LIFX smart bulbs}
\description{
The lifx R package is an interface to the [lifx smart bulb api](https://api.developer.lifx.com/docs).
}
\section{most important lifx functions}{

\itemize{
 \item{ \code{\link{lx_save_token}}}
 \item{ \code{\link{lx_list_lights}}}
 \item{ \code{\link{lx_color}}}
 \item{ \code{\link{lx_effect_breathe}} and other effects}
}
}

