% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{lx_check_color}
\alias{lx_check_color}
\title{check if lifx color name is valid}
\usage{
lx_check_color(color_string, token = get_lifx_token())
}
\arguments{
\item{color_string}{a color string in lifx api format (can be made with lx_color_name() )}
}
\description{
check if lifx color name is valid
}
