% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stuff.R
\name{lx_color}
\alias{lx_color}
\title{picking a color by name or hsbk}
\usage{
lx_color(name = NULL, hue = NULL, saturation = NULL,
  brightness = NULL, kelvin = NULL, token = get_lifx_token())
}
\arguments{
\item{name}{a color name, hexcode or string in lifx api format (see https://api.developer.lifx.com/docs/colors). If this parameter is used, all others will be ignored.}

\item{hue}{set the hue (0-255)}

\item{saturation}{set the saturation (0-1)}

\item{brightness}{set the brightness (0-1)}

\item{kelvin}{set the colour temperature. limits depend on the specific lamp; limits are likely in the range of 2500-9000}
}
\description{
picking a color by name or hsbk
}
