% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx_endpoints.R
\name{lx_delta}
\alias{lx_delta}
\title{Change light state relative to current state (wrapper for POST state delta}
\usage{
lx_delta(
  hue = NULL,
  saturation = NULL,
  brightness = NULL,
  kelvin = NULL,
  infrared = NULL,
  duration = 0,
  power = NULL,
  selector = "all",
  token = get_lifx_token()
)
}
\arguments{
\item{hue}{set the hue (0-255)}

\item{saturation}{set the saturation (0-1)}

\item{brightness}{set the brightness (0-1)}

\item{kelvin}{set the colour temperature. limits depend on the specific lamp; limits are likely in the range of 2500-9000}

\item{infrared}{infrared brightness (0-1)}

\item{duration}{in seconds, how long to perform the transition}

\item{power}{character string - if set to "on", turns the light on, if set to "off" turns it off.}

\item{selector}{lifx api "selector" such as "all", "id:12345", or "location:kitchen". Can be created with \code{\link{lx_selector}} or written manually (see \url{https://api.developer.lifx.com/docs/selectors}}

\item{token}{API token (see ?save_lifx_token). If left empty, the token is retreived from the environmental variable if available. (see \code{\link{save_lifx_token}})}
}
\description{
Change light state relative to current state (wrapper for POST state delta
}
\references{
\url{https://api.developer.lifx.com/docs/state-delta}
}
