% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{lx_effect}
\alias{lx_effect}
\title{"Breathe" effect}
\usage{
lx_effect(
  effect,
  color,
  from_color = NULL,
  period = NULL,
  cycles = NULL,
  persist = NULL,
  power_on = NULL,
  peak = 0.5,
  selector = "all",
  token = get_lifx_token()
)
}
\arguments{
\item{color}{color The color to use for the breathe effect. use lx_color() as input}

\item{from_color}{The color to start the effect from. If this parameter is omitted then the color the bulb is currently set to is used instead.}

\item{period}{The time in seconds for one cycle of the effect.}

\item{cycles}{The number of times to repeat the effect.}

\item{persist}{boolean; If FALSE set the light back to its previous value when effect ends, if true leave the last effect color.}

\item{power_on}{boolean; If true, turn the bulb on if it is not alxlready on.}

\item{peak}{Defines where in a period the target color is at its maximum. Minimum 0.0, maximum 1.0.}
}
\description{
"Breathe" effect
}
