% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{lx_effect_breathe}
\alias{lx_effect_breathe}
\title{"Breathe" effect}
\usage{
lx_effect_breathe(
  color,
  from_color = NULL,
  period = NULL,
  cycles = NULL,
  persist = NULL,
  power_on = NULL,
  peak = 0.5,
  selector = "all",
  token = get_lifx_token()
)
}
\arguments{
\item{color}{color The color to use for the breathe effect. use lx_color() as input}

\item{from_color}{The color to start the effect from. If this parameter is omitted then the color the bulb is currently set to is used instead.}

\item{period}{The time in seconds for one cycle of the effect.}

\item{cycles}{The number of times to repeat the effect.}

\item{persist}{boolean; If FALSE set the light back to its previous value when effect ends, if true leave the last effect color.}

\item{power_on}{If FALSE, does not turn light on if it is off}

\item{peak}{Defines where in a period the target color is at its maximum. Minimum 0.0, maximum 1.0.}

\item{selector}{lifx api "selector" such as "all", "id:12345", or "location:kitchen". Can be created with \code{\link{lx_selector}} or written manually (see \url{https://api.developer.lifx.com/docs/selectors}}

\item{token}{API token (see ?save_lifx_token). If left empty, the token is retreived from the environmental variable if available. (see \code{\link{save_lifx_token}})}
}
\description{
"Breathe" effect
}
