% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr_wrappers.R
\name{lx_GET}
\alias{lx_GET}
\title{GET request}
\usage{
lx_GET(selector = "all", endpoint, token = get_lifx_token())
}
\arguments{
\item{selector}{lifx api "selector" such as "all", "id:12345", or "location:kitchen". Can be created with \code{\link{lx_selector}} or written manually (see \url{https://api.developer.lifx.com/docs/selectors}}

\item{endpoint}{the API endpoint to call; basically the last part of the API url after the light selector}

\item{token}{the API token; see ?lx_auth}
}
\description{
GET request
}
