% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light_color.R
\name{lx_light_color}
\alias{lx_light_color}
\title{we wrap in three steps:}
\usage{
lx_light_color(hue = NULL, saturation = NULL, brightness = NULL,
  kelvin = NULL, duration = NULL, infrared = NULL,
  color_name = NULL, fast = NULL, selector = "all", ...,
  token = get_lifx_token())
}
\arguments{
\item{hue}{set the hue (0-255)}

\item{saturation}{set the saturation (0-1)}

\item{brightness}{set the brightness (0-1)}

\item{kelvin}{set the colour temperature. limits depend on the specific lamp; limits are likely in the range of 2500-9000}

\item{duration}{in seconds, how long to perform the transition}

\item{color_name}{a color name, hexcode or output from lx_color() (in lifx api format (see https://api.developer.lifx.com/docs/colors). If this parameter is used, other parameters may be ignored.}

\item{fast}{Execute the query fast, without initial state checks and wait for no results. See https://api.developer.lifx.com/docs/set-state}

\item{selector}{lifx API selector criteria; can be created with lx_select() or written manually}

\item{...}{further parameters passed to lx_state()}

\item{token}{API token (see ?get_lifx_token)}
}
\description{
1. a user facing function based on what the user probably wants to do (i.e. set light color)
2. a function wrapping each specific api endpoint (i.e. lx_state())
3. a function wrapping httr verbs PUT POST GET
}
