% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{lx_list_lights}
\alias{lx_list_lights}
\title{list available lights}
\usage{
lx_list_lights(selector = "all", token = get_lifx_token())
}
\arguments{
\item{selector}{lifx api "selector" such as "all", "id:12345", or "location:kitchen". Can be created with \code{\link{lx_selector}} or written manually (see \url{https://api.developer.lifx.com/docs/selectors}}

\item{token}{API token (see ?save_lifx_token). If left empty, the token is retreived from the environmental variable if available. (see \code{\link{save_lifx_token}})}
}
\value{
the API response
}
\description{
list available lights
}
