% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr_wrappers.R
\name{lx_PUT}
\alias{lx_PUT}
\title{PUT request}
\usage{
lx_PUT(selector = "all", endpoint, token, ...)
}
\arguments{
\item{selector}{lifx api "selector" such as "all", "id:12345", or "location:kitchen" - see https://api.developer.lifx.com/docs/selectors}

\item{endpoint}{the API endpoint to call; basically the last part of the API url after the light selector}

\item{token}{the API token; see ?lx_auth}

\item{...}{named values to add to the request body}
}
\description{
PUT request
}
