% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_limit.R
\name{lx_rate_limit}
\alias{lx_rate_limit}
\title{get lifx API rate limit}
\usage{
lx_rate_limit(token = get_lifx_token())
}
\arguments{
\item{token}{API token (see ?save_lifx_token). If left empty, the token is retrieved from the environmental variable if available. (see \code{\link{save_lifx_token}})}
}
\value{
a named vector of three numbers:

1. `limit`: The rate limit
2. `remaining`: how many calls are remaining
3. `reset`: the Unix timestamp for when the next window begins. Usually every minute.
}
\description{
get lifx API rate limit
}
