% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr_wrappers.R
\name{get_lifx_token}
\alias{get_lifx_token}
\title{retrieve lifx_token from R environment}
\usage{
get_lifx_token()
}
\description{
#' @details To use the lifx API, you need to get a personal access token from your lifx account. Usually you save API tokens in your r environment file; that way you only have to enter it once per system. Once you have your token, you can use `save_lifx_token()` to do that.
How to get a token:
1. go to \url{https://cloud.lifx.com/sign_in} and sign in (if you do not have an account, you must download the mobile app and register there.
2. generate or look up your access token
}
\details{
You do not need to save the token in the environment; you can use all functions in this package by passing a valid `token` argument.
}
\seealso{
\code{\link{has_lifx_token}}, \code{\link{save_lifx_token}}
}
