% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_name.R
\name{lx_check_color}
\alias{lx_check_color}
\title{check if lifx color name is valid}
\usage{
lx_check_color(color_name, token = lx_get_token())
}
\arguments{
\item{color_name}{a color string in lifx api format (can be made with \code{\link{lx_color_name}} )}

\item{token}{API token (see ?lx_save_token). If left empty, the token is retrieved from the environmental variable if available. (see \code{\link{lx_save_token}})}
}
\description{
check if lifx color name is valid
}
