% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{lx_check_color}
\alias{lx_check_color}
\title{check if lifx color name is valid}
\usage{
lx_check_color(color_name, token = get_lifx_token())
}
\arguments{
\item{color_name}{a color string in lifx api format (can be made with \code{\link{lx_color_name}} )}

\item{token}{API token (see ?save_lifx_token). If left empty, the token is retreived from the environmental variable if available. (see \code{\link{save_lifx_token}})}
}
\description{
check if lifx color name is valid
}
