% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{lx_effect_off}
\alias{lx_effect_off}
\title{Turn effects off}
\usage{
lx_effect_off(power_off = FALSE, selector = "all", token = lx_get_token())
}
\arguments{
\item{power_off}{If TRUE, also turns the light(s) off}

\item{selector}{'LIFX' api "selector" such as "all", "id:12345", or "location:kitchen". Can be created with \code{\link{lx_selector}} or written manually (see \url{https://api.developer.lifx.com/docs/selectors}}

\item{token}{API token (see ?lx_save_token). If left empty, the token is retrieved from the environmental variable if available. (see \code{\link{lx_save_token}})}
}
\value{
an 'httr' response object (see \code{\link[httr]{response}})
}
\description{
Turn effects off
}
\examples{
\dontrun{lx_effect_off()}
}
