% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_lights.R
\name{lx_list_lights}
\alias{lx_list_lights}
\title{list available lights}
\usage{
lx_list_lights(selector = "all", token = lx_get_token())
}
\arguments{
\item{selector}{'LIFX' api "selector" such as "all", "id:12345", or "location:kitchen". Can be created with \code{\link{lx_selector}} or written manually (see \url{https://api.developer.lifx.com/docs/selectors}}

\item{token}{API token (see ?lx_save_token). If left empty, the token is retrieved from the environmental variable if available. (see \code{\link{lx_save_token}})}
}
\value{
an 'httr' response object (see \code{\link[httr]{response}})
}
\description{
list available lights
}
\examples{
\dontrun{
lx_list_lights()
lx_list_lights(
  lx_selector(location = "kitchen")
)
}
}
