% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{lx_list_lights}
\alias{lx_list_lights}
\title{list available lights}
\usage{
lx_list_lights(selector = "all", token = get_lifx_token())
}
\arguments{
\item{selector}{which lights to look for?}
}
\description{
list available lights
}
