% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light_color.R
\name{lx_state}
\alias{lx_state}
\title{set light state (wrapper for PUT Set State https://api.developer.lifx.com/docs/set-state)}
\usage{
lx_state(power = NULL, color = NULL, brightness = NULL,
  infrared = NULL, duration = 0, fast = FALSE, selector = "all",
  token = get_lifx_token())
}
\arguments{
\item{power}{"on" or "off"}

\item{additive}{if TRUE, values will be added to current values (ie. brightness=0.1 will make the lamp 0.1 ligther than currenly). 'color' argument only works if additive = FALSE (default).}
}
\description{
set light state (wrapper for PUT Set State https://api.developer.lifx.com/docs/set-state)
}
