% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light_toggle.R
\name{lx_toggle}
\alias{lx_toggle}
\title{Toggle light}
\usage{
lx_toggle(duration = 0, selector = "all", token = get_lifx_token())
}
\arguments{
\item{duration}{in seconds, how long to perform the transition}

\item{select}{lifx API selector criteria (see ?lx_select)}
}
\description{
Toggle light
}
