% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle.R
\name{lx_toggle}
\alias{lx_toggle}
\title{Toggle light}
\usage{
lx_toggle(duration = 0, selector = "all", token = lx_get_token())
}
\arguments{
\item{duration}{in seconds, how long to perform the transition}

\item{selector}{'LIFX' api "selector" such as "all", "id:12345", or "location:kitchen". Can be created with \code{\link{lx_selector}} or written manually (see \url{https://api.developer.lifx.com/docs/selectors}}

\item{token}{API token (see ?lx_save_token). If left empty, the token is retrieved from the environmental variable if available. (see \code{\link{lx_save_token}})}
}
\description{
Toggle light
}
\examples{
\dontrun{
lx_toggle(duration = 5)
lx_toggle(selector = lx_selector(location = "kitchen"))
}
}
